requirements.txt
Django==6.0.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.8.2
dj-database-url==2.3.0
